// Frontend/react-movies/src/movies/movies.model.d.ts
import { actorMovieDTO } from "../actors/actors.model";

export interface movieDTO{
 id:number;
 title:string;
 poster:string;
}

export interface movieCreationDTO{
 title:string;
 inTheaters:boolean;
 trailer:string;
 releaseDate?:Date;
 poster?:File;
 posterURL?:string;
 genresIds?:number[];
 movieTheatersIds?:number[];
 actors?:actorMovieDTO[];
}

export interface landingPageDTO{
 intheatres?:movieDTO[];
 upcomingReleases?:movieDTO[];

}